fixed styles for fees page
