+++ packages/social-trader/src/routes/ssr/landing-page/components/fees-info/fees-info.scss
@@ -116,21 +116,26 @@
     .fees-table__cell {
       font-size: 14px;
       font-weight: 400;
+      border-bottom: none;
     }
   }
 
   &__cell {
     min-width: 50px;
-    padding: 18px 0;
+    padding: 18px 0 18px 20px;
+    border-bottom: 1px solid $landing-bg-gray;
 
     &:first-child {
       font-weight: 600;
-      padding-left: 20px;
       width: 200px;
     }
 
     &--colspan {
       text-align: center;
+    }
+
+    &--without-border {
+      border-bottom: none;
     }
 
     &--color-primary {
