+++ packages/social-trader/src/routes/ssr/landing-page/components/fees-info/fees-trading.tsx
@@ -1,7 +1,7 @@
 import "./fees-info.scss";
 
-import classNames from "classnames";
 import React from "react";
+import FeesTradingDiscount from "routes/ssr/landing-page/components/fees-info/fees-trading-discount";
 
 interface Props {
   className?: string;
@@ -114,7 +114,9 @@ const _FeesTrading: React.FC<Props> = ({ className }) => (
             </thead>
             <tbody>
               <tr className="fees-table__row">
-                <td className="fees-table__cell">Forex</td>
+                <td className="fees-table__cell fees-table__cell--without-border">
+                  Forex
+                </td>
                 <td className="fees-table__cell" rowSpan={4}>
                   for 100 000 USD
                 </td>
@@ -141,10 +143,14 @@ const _FeesTrading: React.FC<Props> = ({ className }) => (
                 </td>
               </tr>
               <tr className="fees-table__row">
-                <td className="fees-table__cell">Spot metals</td>
+                <td className="fees-table__cell fees-table__cell--without-border">
+                  Spot metals
+                </td>
               </tr>
               <tr className="fees-table__row">
-                <td className="fees-table__cell">Index</td>
+                <td className="fees-table__cell fees-table__cell--without-border">
+                  Index
+                </td>
               </tr>
               <tr className="fees-table__row">
                 <td className="fees-table__cell">Spot commodities</td>
@@ -183,7 +189,9 @@ const _FeesTrading: React.FC<Props> = ({ className }) => (
                 <td className="fees-table__cell" />
               </tr>
               <tr className="fees-table__row">
-                <td className="fees-table__cell">UK</td>
+                <td className="fees-table__cell fees-table__cell--without-border">
+                  UK
+                </td>
                 <td className="fees-table__cell" rowSpan={6}>
                   order volume (%)
                 </td>
@@ -196,7 +204,9 @@ const _FeesTrading: React.FC<Props> = ({ className }) => (
                 <td className="fees-table__cell">0.18</td>
               </tr>
               <tr className="fees-table__row">
-                <td className="fees-table__cell">Germane</td>
+                <td className="fees-table__cell fees-table__cell--without-border">
+                  Germane
+                </td>
                 <td className="fees-table__cell">1.00</td>
                 <td className="fees-table__cell">0.70</td>
                 <td className="fees-table__cell">0.65</td>
@@ -206,7 +216,9 @@ const _FeesTrading: React.FC<Props> = ({ className }) => (
                 <td className="fees-table__cell">0.45</td>
               </tr>
               <tr className="fees-table__row">
-                <td className="fees-table__cell">Switzerland</td>
+                <td className="fees-table__cell fees-table__cell--without-border">
+                  Switzerland
+                </td>
                 <td className="fees-table__cell">0.60</td>
                 <td className="fees-table__cell">0.42</td>
                 <td className="fees-table__cell">0.39</td>
@@ -216,7 +228,9 @@ const _FeesTrading: React.FC<Props> = ({ className }) => (
                 <td className="fees-table__cell">0.27</td>
               </tr>
               <tr className="fees-table__row">
-                <td className="fees-table__cell">France</td>
+                <td className="fees-table__cell fees-table__cell--without-border">
+                  France
+                </td>
                 <td className="fees-table__cell">0.40</td>
                 <td className="fees-table__cell">0.28</td>
                 <td className="fees-table__cell">0.26</td>
@@ -226,7 +240,9 @@ const _FeesTrading: React.FC<Props> = ({ className }) => (
                 <td className="fees-table__cell">0.18</td>
               </tr>
               <tr className="fees-table__row">
-                <td className="fees-table__cell">Spain</td>
+                <td className="fees-table__cell fees-table__cell--without-border">
+                  Spain
+                </td>
                 <td className="fees-table__cell">0.80</td>
                 <td className="fees-table__cell">0.56</td>
                 <td className="fees-table__cell">0.52</td>
@@ -268,7 +284,9 @@ const _FeesTrading: React.FC<Props> = ({ className }) => (
                 <td className="fees-table__cell" />
               </tr>
               <tr className="fees-table__row">
-                <td className="fees-table__cell">Hong Kong</td>
+                <td className="fees-table__cell fees-table__cell--without-border">
+                  Hong Kong
+                </td>
                 <td className="fees-table__cell" rowSpan={6}>
                   order volume (%)
                 </td>
@@ -311,101 +329,6 @@ const _FeesTrading: React.FC<Props> = ({ className }) => (
     </div>
   </div>
 );
-
-export const FeesTradingDiscount: React.FC<{
-  dark?: boolean;
-  white?: boolean;
-}> = ({ dark, white }) => {
-  return (
-    <>
-      <div className="fees-info__notes">
-        <h3>To be eligible for a trading discount</h3>
-        <ul
-          className={classNames("fees-info__list-notes", {
-            "fees-info__list-notes--dark": dark
-          })}
-        >
-          <li className="fees-info__note-item">
-            There must be more than 1 GVT stored in the wallet
-          </li>
-          <li className="fees-info__note-item">
-            The "Using GVT to pay for fees" function must be enabled. The fee
-            will be charged at a discount depending on the amount of GVT stored
-            in the wallet.
-          </li>
-          <li className="fees-info__note-item">
-            The discounted fee will be charged from the GVT wallet, not from the
-            trading account.
-          </li>
-        </ul>
-      </div>
-      <div className="fees-info__notes">
-        <h4>Conditions for obtaining discounts</h4>
-        <div className="fees-info__table-wrapper">
-          <table
-            className={classNames("fees-table", {
-              "fees-table--dark": dark,
-              "fees-table--white-head": white
-            })}
-          >
-            <thead className="fees-table__head">
-              <tr className="fees-table__row">
-                <th className="fees-table__cell fees-table__cell--width-bg">
-                  Wallet
-                </th>
-                <th className="fees-table__cell">Amount</th>
-              </tr>
-            </thead>
-            <tbody>
-              <tr className="fees-table__row">
-                <td className="fees-table__cell">From 1 GVT up to 10</td>
-                <td className="fees-table__cell">30%</td>
-              </tr>
-              <tr className="fees-table__row">
-                <td className="fees-table__cell">Up to 25 GVT</td>
-                <td className="fees-table__cell">35%</td>
-              </tr>
-              <tr className="fees-table__row">
-                <td className="fees-table__cell">Up to 50 GVT</td>
-                <td className="fees-table__cell">40%</td>
-              </tr>
-              <tr className="fees-table__row">
-                <td className="fees-table__cell">Up to 100 GVT</td>
-                <td className="fees-table__cell">45%</td>
-              </tr>
-              <tr className="fees-table__row">
-                <td className="fees-table__cell">Up to 500 GVT</td>
-                <td className="fees-table__cell">50%</td>
-              </tr>
-              <tr className="fees-table__row">
-                <td className="fees-table__cell">500 GVT+</td>
-                <td className="fees-table__cell">55%</td>
-              </tr>
-            </tbody>
-          </table>
-        </div>
-      </div>
-      <div className="fees-info__notes">
-        <h3>Note</h3>
-        <ul
-          className={classNames("fees-info__list-notes", {
-            "fees-info__list-notes--dark": dark
-          })}
-        >
-          <li className="fees-info__note-item">
-            If you switch this function off, the fee will become "Regular" and
-            charge 100%.
-          </li>
-          <li className="fees-info__note-item">
-            If you have enabled the "Using GVT to pay for fees" function but
-            have an insufficient amount of GVT stored in the wallet, this
-            function will no longer work and the fee will become Regular.
-          </li>
-        </ul>
-      </div>
-    </>
-  );
-};
 
 const FeesTrading = React.memo(_FeesTrading);
 export default FeesTrading;
